Edit-Navbar admin
