+++ src/app/admin/layout.js
@@ -3,7 +3,8 @@
 import { usePathname } from 'next/navigation'
 import UserAuthContextProvider from "@/context/UserAuthContext";
 import ProtectedRoute from "@/auth/ProtectedRoute";
-import ToastNotification from '@/components/toastNotification';
+import ToastNotification from '@/components/ToastNotification';
+import Navbar from '@/components/nav_admin';
 
 export default function AdminLayout({ children }) {
   const pathname = usePathname();
@@ -13,8 +14,9 @@ export default function AdminLayout({ children }) {
       <div className="min-h-screen bg-gray-100">
         {/* เพิ่ม Navbar หรือ Sidebar ของ Admin */}
         <ToastNotification/>
-        {isPublic ? children : <ProtectedRoute>{children}</ProtectedRoute>}
+        {isPublic ? children : <ProtectedRoute><Navbar/>{children}</ProtectedRoute>}
       </div>
     </UserAuthContextProvider>
   );
 }
+
