+++ src/app/admin/page.js
@@ -3,16 +3,13 @@ import React, { useState } from "react"
 import { useUserAuth } from "@/context/UserAuthContext";
 import { useRouter } from 'next/navigation'
 
+
 export default function AdminHome() {
   const router = useRouter()
   
   const { Logout,user } = useUserAuth()
 
   console.log(user)
-
-  const handleManageUsers = () => {
-    router.push('/admin/member')
-  };
 
   const handleLogout = async () => {
     try {
@@ -31,10 +28,6 @@ export default function AdminHome() {
       <h1 className="text-3xl font-bold text-gray-800 text-center">Welcome to Admin Dashboard</h1>
 
       <p className="text-gray-800">Hi, {user?.email}</p>
-
-      <button onClick={handleManageUsers} className="px-5 py-2.5 bg-blue-500 hover:bg-blue-600 text-white rounded-xl shadow-md transition duration-200 cursor-pointer">
-        จัดการทำเนียบสภา
-      </button>
 
       <div className="flex flex-row space-x-4">
         <button onClick={handleLogout} className="px-5 py-2.5 bg-red-500 hover:bg-red-600 text-white rounded-xl shadow-md transition duration-200 cursor-pointer">
